Lab 10/requirements.txt
flask==3.0.3
groq==0.13.0
python-dotenv==1.0.1